+++ src/connectedmachine/azext_connectedmachine/aaz/latest/connectedmachine/license/__init__.py
@@ -12,6 +12,8 @@ from .__cmd_group import *
 from ._create import *
 from ._delete import *
 from ._list import *
+from ._list_by_rg import *
 from ._show import *
 from ._update import *
+from ._validate import *
 from ._wait import *